Fix pytest configuration for retry logic in integration and docs tests
